// backend/src/modules/task/core/events/TaskArchived.d.ts
import DomainEvent from '../../../shared/core/events/DomainEvent';
import type DateTime from '../../../shared/core/objects/DateTime';
import type IdEntity from '../../../shared/core/objects/IdEntity';
export default class TaskArchived extends DomainEvent {
    constructor(key: string, date: DateTime, actor: IdEntity, idProject: IdEntity, idEntity: IdEntity);
}
//# sourceMappingURL=TaskArchived.d.ts.map